requirements.txt
certifi==2019.9.11
cffi==1.13.1
Click==7.0
cryptography==2.8
Flask==1.1.1
future==0.18.1
invoke==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
pycparser==2.19
python-telegram-bot==12.2.0
six==1.12.0
tornado==6.0.3
Werkzeug==0.16.0
